The branch outgrew its name a long way. What landed:

Cosmetics and weapons draw on players. They are separate entities carrying no
origin and, for most, no ordinary model index: the parent is moveparent rather
than m_hOwnerEntity (taking the owner claimed 220 syringe projectiles as worn
items), the handle is the low 11 bits with the invalid test applied to the whole
value first, and the model lives in the DynamicModels string table at
(-2 - index) >> 1 for even dynamic indices. Bone merge copies the wearer's
matrices for bones matched BY NAME and builds the rest from the item's own
hierarchy -- skipping that second half left a gibus on the floor.

Players stand up. Pose parameters are stored normalised, a sequence names a grid
of animations blended three ways with a normalised lerp after QuaternionAlign,
and the bug that actually laid everyone down was a substring sequence lookup
returning AttackStand_PRIMARY for Stand_PRIMARY.

A free camera ported from AngleVectors and MatrixBuildPerspective, asserted on
where points land rather than on matrix entries -- which caught a mirrored world
twice, since Source's +Y points left and AngleVectors hands back right. Having a
camera that moves then exposed three shortcuts taken for the fixed one: a
Normal.Z back-face cull, a depth bias standing in for decal placement, and a
pass with depth writes left off.

Overlays clip against the faces they name, Sutherland-Hodgman, which is what
that face list is for. It is not a set of candidate surfaces, and reading it as
one is why four renderer theories died before the BSP was asked.

Bodygroups are selected per entity at draw time, tagged (part, model) through
the batches, because selecting at load desynchronises the .vtx strip groups.

And the test suite stopped fighting the machine. It was sending synthesized
input, which goes to whichever window holds the foreground, so runs typed into
the tester's browser; everything reachable through UI Automation now goes that
way, and the keys full screen genuinely needs are guarded -- foreground taken,
arrival verified, refused rather than sent blind. One viewer serves the whole
assembly instead of four that accumulated until they locked the executable. The
whole solution runs in one command: 1599 tests, 0 failed.

Three application defects came out of that work rather than out of the renderer:
full screen lost keyboard focus and could only be escaped by alt-tabbing, the
transport bar told a screen reader two different wordings for its speed, and the
screenshot had no route but F12 -- nothing a keyboard-only user could reach. A
fourth came out of running everything at once: ScenePlayer.Yaw was never carried,
so every player in a frame faced due east.

Known open, all in RISKS: B73 (bodygroups correct at model, demo and packing and
still drawing '?', narrowed to the one unmeasured hop), B74, B71 rolling doors,
B70, B76.

+++ managed/Tf2DemoSalvage.Core/Primitives/Snappy.cs
@@ -195,7 +195,9 @@ public static class Snappy
         length++;
         Need(compressed, read, length);
 
-        if (written + length > output.Length)
+        // Long for the same reason as Need: written plus a length near int.MaxValue wraps
+        // negative, and a negative is never greater than the output's length.
+        if ((long)written + length > output.Length)
         {
             throw TooLong(output.Length);
         }
@@ -259,7 +261,16 @@ public static class Snappy
 
     private static void Need(ReadOnlySpan<byte> compressed, int read, int count)
     {
-        if (read + count > compressed.Length)
+        // **Widened to long, because the sum is what overflowed.** A length read from the stream
+        // can be int.MaxValue, and int.MaxValue plus a read offset wraps to a negative - which is
+        // not greater than the buffer length, so this guard passed and the value went on to die
+        // in Slice's argument validation instead.
+        //
+        // Every guard here is written as "is this number too large", and that shape cannot catch
+        // a number that becomes small by wrapping. Doing the comparison in a type the sum fits in
+        // is the fix, and it belongs here rather than at any one call site: this is the check all
+        // of them share.
+        if ((long)read + count > compressed.Length)
         {
             throw new InvalidDataException(string.Create(
                 CultureInfo.InvariantCulture,
